+++ src/MarginTrading.Common/BackendContracts/ChangeOrderLimitsBackendRequest.cs
@@ -5,8 +5,8 @@
         public string ClientId { get; set; }
         public string OrderId { get; set; }
         public string AccountId { get; set; }
-        public double TakeProfit { get; set; }
-        public double StopLoss { get; set; }
-        public double ExpectedOpenPrice { get; set; }
+        public decimal TakeProfit { get; set; }
+        public decimal StopLoss { get; set; }
+        public decimal ExpectedOpenPrice { get; set; }
     }
 }
