+++ src/MarginTrading.Common/ClientContracts/ChangeOrderLimitsClientRequest.cs
@@ -14,8 +14,8 @@ namespace MarginTrading.Common.ClientContracts
         public string OrderId { get; set; }
         [Required]
         public string AccountId { get; set; }
-        public double TakeProfit { get; set; }
-        public double StopLoss { get; set; }
-        public double ExpectedOpenPrice { get; set; }
+        public decimal TakeProfit { get; set; }
+        public decimal StopLoss { get; set; }
+        public decimal ExpectedOpenPrice { get; set; }
     }
 }
