+++ src/MarginTrading.Common/ClientContracts/DepositWithdrawClientRequest.cs
@@ -10,6 +10,6 @@ namespace MarginTrading.Common.ClientContracts
         [Required]
         public string AccountId { get; set; }
         [Required]
-        public double? Volume { get; set; }
+        public decimal? Volume { get; set; }
     }
 }
