+++ src/MarginTrading.Core/IFplService.cs
@@ -6,6 +6,6 @@ namespace MarginTrading.Core
     public interface IFplService
     {
         void UpdateOrderFpl(IOrder order, FplData fplData);
-        double GetMatchedOrdersPrice(List<MatchedOrder> matchedOrders, string instrument);
+        decimal GetMatchedOrdersPrice(List<MatchedOrder> matchedOrders, string instrument);
     }
 }
